subject_matters_rag/requirements.txt
streamlit==1.29.0
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.2
psycopg2-binary==2.9.9
pgvector==0.2.4
neo4j==5.14.0
pypdf2==3.0.1
sentence-transformers==2.2.2
tiktoken==0.5.1
python-dotenv==1.0.0
numpy==1.24.3